New Features
New toggle for setting if the worm can only move if it is eating
something.
New toggle for setting if recipe matches must be in order
New toggle for displaying how many moves the worm has taken.
New field to set how many moves the worm can take before it is killed.
If set to 0, the worm will not die from moving.